+++ nbproject/private/private.xml
@@ -3,9 +3,8 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/F:/Git_Bash/freestyle/src/cn/edu/pku/datasource/FTranslatePoint.java</file>
             <file>file:/F:/Git_Bash/freestyle/src/cn/edu/pku/datasource/SimpleFeatureManager.java</file>
-            <file>file:/F:/Git_Bash/freestyle/src/cn/edu/pku/datasource/ShapefileManager.java</file>
+            <file>file:/F:/Git_Bash/freestyle/src/cn/edu/pku/datasource/GeometryManager.java</file>
         </group>
     </open-files>
 </project-private>
