+++ src/cn/edu/pku/datasource/SimpleFeatureManager.java
@@ -5,6 +5,7 @@
  */
 package cn.edu.pku.datasource;
 
+import static cn.edu.pku.datasource.GeometryManager.createOnePoint;
 import com.vividsolutions.jts.geom.Coordinate;
 import com.vividsolutions.jts.geom.CoordinateSequence;
 import com.vividsolutions.jts.geom.Geometry;
@@ -109,153 +110,7 @@ public class SimpleFeatureManager {
         return sGeometryList;
     }
 
-    /**
-     * 根据经纬度生成点
-     *
-     * @param latitude 纬度
-     * @param longitude 经度
-     * @return 点对象
-     */
-    public static Point createOnePoint(double latitude, double longitude) {
-        GeometryFactory geometryFactory = new GeometryFactory();
-        Point point = geometryFactory.createPoint(new Coordinate(longitude, latitude));
-        return point;
-    }
-
-    /**
-     * 根据经纬度生成线
-     *
-     * @param latArray 纬度序列
-     * @param lngArray 经度序列
-     * @return 线对象
-     * @throws Exception
-     */
-    public static LineString createOneLineString(double[] latArray, double[] lngArray) throws Exception {
-        GeometryFactory geometryFactory = new GeometryFactory();
-        int latLength = latArray.length;
-        int lngLength = lngArray.length;
-        if (latLength != lngLength) {
-            throw new Exception("数据长度不匹配");
-        }
-        ArrayList<Coordinate> coordinateList = new ArrayList<Coordinate>();
-        for (int i = 0; i < latLength; i++) {
-            coordinateList.add(new Coordinate(lngArray[i], latArray[i]));
-        }
-        Coordinate[] coorArray = new Coordinate[coordinateList.size()];
-        LineString lineString = geometryFactory.createLineString(coordinateList.toArray(coorArray));
-        return lineString;
-    }
-
-    /**
-     * 根据坐标点生成线
-     *
-     * @param coordinateArray 坐标点序列
-     * @return 线对象
-     * @throws Exception
-     */
-    public static LineString createOneLineString(Coordinate[] coordinateArray) throws Exception {
-        GeometryFactory geometryFactory = new GeometryFactory();
-        LineString lineString = geometryFactory.createLineString(coordinateArray);
-        return lineString;
-    }
-
-    /**
-     * 根据经纬度序列生成环
-     *
-     * @param latArray 纬度序列
-     * @param lngArray 经度序列
-     * @return 环对象
-     * @throws Exception
-     */
-    public static LinearRing createOneLinearRing(double[] latArray, double[] lngArray) throws Exception {
-        GeometryFactory geometryFactory = new GeometryFactory();
-        int latLength = latArray.length;
-        int lngLength = lngArray.length;
-        if (latLength != lngLength) {
-            throw new Exception("数据长度不匹配");
-        }
-        ArrayList<Coordinate> coordinateList = new ArrayList<Coordinate>();
-        for (int i = 0; i < latLength; i++) {
-            coordinateList.add(new Coordinate(lngArray[i], latArray[i]));
-        }
-        Coordinate[] coorArray = new Coordinate[coordinateList.size()];
-        LinearRing linearRing = geometryFactory.createLinearRing(coordinateList.toArray(coorArray));
-        return linearRing;
-    }
-
-    /**
-     * 根据坐标点序列生成环
-     *
-     * @param coordinateArray 坐标点序列
-     * @return 环对象
-     * @throws Exception
-     */
-    public static LinearRing createOneLinearRing(Coordinate[] coordinateArray) throws Exception {
-        GeometryFactory geometryFactory = new GeometryFactory();
-        LinearRing linearRing = geometryFactory.createLinearRing(coordinateArray);
-        return linearRing;
-    }
-
-    /**
-     * 根据坐标点序列生成多边形
-     *
-     * @param latArray 纬度序列
-     * @param lngArray 经度序列
-     * @return 多边形对象
-     * @throws Exception
-     */
-    public static Polygon createOnePolygon(double[] latArray, double[] lngArray) throws Exception {
-        GeometryFactory geometryFactory = new GeometryFactory();
-        int latLength = latArray.length;
-        int lngLength = lngArray.length;
-        if (latLength != lngLength) {
-            throw new Exception("数据长度不匹配");
-        }
-        ArrayList<Coordinate> coordinateList = new ArrayList<Coordinate>();
-        for (int i = 0; i < latLength; i++) {
-            coordinateList.add(new Coordinate(lngArray[i], latArray[i]));
-        }
-        Coordinate[] coorArray = new Coordinate[coordinateList.size()];
-        Polygon polygon = geometryFactory.createPolygon(coordinateList.toArray(coorArray));
-        return polygon;
-    }
-
-    /**
-     * 根据坐标点序列生成多边形
-     *
-     * @param coordinateArray 坐标点序列
-     * @return 多边形对象
-     * @throws Exception
-     */
-    public static Polygon createOnePolygon(Coordinate[] coordinateArray) throws Exception {
-        GeometryFactory geometryFactory = new GeometryFactory();
-        Polygon polygon = geometryFactory.createPolygon(coordinateArray);
-        return polygon;
-    }
-
-    /**
-     * 根据输入生成一个圆
-     *
-     * @param x 横坐标
-     * @param y 纵坐标
-     * @param RADIUS 半径
-     * @param SIDES 圆上点的个数
-     * @return
-     */
-    public static Polygon createOneCircle(double x, double y, final double RADIUS, int SIDES) {
-        Coordinate coords[] = new Coordinate[SIDES + 1];
-        for (int i = 0; i < SIDES; i++) {
-            double angle = ((double) i / (double) SIDES) * Math.PI * 2.0;
-            double dx = Math.cos(angle) * RADIUS;
-            double dy = Math.sin(angle) * RADIUS;
-            coords[i] = new Coordinate((double) x + dx, (double) y + dy);
-        }
-        coords[SIDES] = coords[0];
-        GeometryFactory geometryFactory = new GeometryFactory();
-        LinearRing ring = geometryFactory.createLinearRing(coords);
-        Polygon polygon = geometryFactory.createPolygon(ring, null);
-        return polygon;
-    }
+    
 
     /**
      * 新建一个点要素
